Removed wrapped call_tool from session and test connection from the base connector + fix tests
